requirements.txt
boto3
faker
confluent-kafka
elasticsearch
apache-airflow
